# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Implements the Mailio server based on https://mirs.mail.io/ specifications",
        "title": "Mailio Server API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/credentials/list/{address}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retruns a list of VCs by mailio address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifiable Credentials"
                ],
                "summary": "List all VCs for a specific mailio address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mailio address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit of VCs to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page token",
                        "name": "pageToken",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/did.Document"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "500": {
                        "description": "error creating server did",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
        "/api/v1/credentials/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Returns a single verifiable credential by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifiable Credentials"
                ],
                "summary": "Get verifiable credential by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VC ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/did.VerifiableCredential"
                        }
                    },
                    "404": {
                        "description": "VC not found",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "500": {
                        "description": "error creating server did",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
        "/api/v1/credentials/{requestId}/verify": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Checks if signature is valid and returns a boolean object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifiable Credentials"
                ],
                "summary": "Verify a verifiable credential",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference ID (request ID, could be anything)",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Verifiable credential to verify",
                        "name": "vc",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/did.VerifiableCredential"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.VCValidationResponse"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the HTTP status code",
                    "type": "integer"
                },
                "message": {
                    "description": "Message is the error message",
                    "type": "string"
                }
            }
        },
        "did.AuthorizedApplication": {
            "type": "object",
            "properties": {
                "approvalDate": {
                    "type": "string"
                },
                "domains": {
                    "description": "domains of the auth application: [example.com]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "description": "target application did: did:example:123456789abcdefghi",
                    "type": "string"
                },
                "userPermissions": {
                    "description": "optional list of permissions specific to a target application",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "did.CredentialStatus": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "https://example.edu/status/24\"",
                    "type": "string"
                },
                "type": {
                    "description": "CredentialStatusList2017",
                    "type": "string"
                }
            }
        },
        "did.CredentialSubject": {
            "type": "object",
            "properties": {
                "authorizedApplication": {
                    "$ref": "#/definitions/did.AuthorizedApplication"
                },
                "id": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                }
            }
        },
        "did.DID": {
            "type": "object"
        },
        "did.Document": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "alsoKnownAs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "authentication": {
                    "type": "array",
                    "items": {}
                },
                "id": {
                    "$ref": "#/definitions/did.DID"
                },
                "keyAgreement": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/did.KeyAgreement"
                    }
                },
                "service": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/did.Service"
                    }
                },
                "verificationMethod": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/did.VerificationMethod"
                    }
                }
            }
        },
        "did.KeyAgreement": {
            "type": "object",
            "properties": {
                "controller": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "publicKeyJwk": {
                    "$ref": "#/definitions/did.PublicKeyJwk"
                },
                "publicKeyMultibase": {
                    "type": "string"
                },
                "type": {
                    "description": "usually X25519KeyAgreementKey2019",
                    "type": "string"
                }
            }
        },
        "did.Proof": {
            "type": "object",
            "properties": {
                "challenge": {
                    "description": "prevent replay attacks",
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "domain": {
                    "description": "prevent replay attacks",
                    "type": "string"
                },
                "jws": {
                    "type": "string"
                },
                "proofPurpose": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "verificationMethod": {
                    "type": "string"
                }
            }
        },
        "did.PublicKeyJwk": {
            "type": "object",
            "properties": {
                "key": {}
            }
        },
        "did.Service": {
            "type": "object",
            "properties": {
                "accept": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "routingKeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "serviceEndpoint": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "did.VerifiableCredential": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "credentialStatus": {
                    "$ref": "#/definitions/did.CredentialStatus"
                },
                "credentialSubject": {
                    "$ref": "#/definitions/did.CredentialSubject"
                },
                "id": {
                    "type": "string"
                },
                "issuanceDate": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string"
                },
                "proof": {
                    "$ref": "#/definitions/did.Proof"
                },
                "type": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "did.VerificationMethod": {
            "type": "object",
            "properties": {
                "controller": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "publicKeyJwk": {
                    "$ref": "#/definitions/did.PublicKeyJwk"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "types.VCValidationResponse": {
            "type": "object",
            "properties": {
                "requestId": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}